+++ src/libs/lib/auth.js
@@ -4,7 +4,6 @@ oscAuth.getLoginUrl = function(config) {
   let returnTo = 'returnTo=' + encodeURIComponent(document.location.toString());
   let loginUrl = config.loginUrl || '/oauth/login?{returnTo}';
   loginUrl = loginUrl.replace(/{returnTo}/, returnTo);
-  console.log(loginUrl);
   return loginUrl;
 }