+++ webpack.config.js
@@ -1,28 +1,28 @@
-const webpack = require("webpack");
+const webpack = require('webpack');
 const path = require('path');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const TerserPlugin = require("terser-webpack-plugin"); 
+const TerserPlugin = require('terser-webpack-plugin'); 
+
 module.exports = {
 
-	devtool: 'source-map',
-	// mode: 'production',
-	mode: 'development',
+	mode: 'production',
 
 	entry: {
-    "all": './src/index.jsx',
-    "base-map": './src/base-map/index.jsx',
-    "button": './src/button/index.jsx',
-    "choices-guide": './src/choices-guide/index.jsx',
-    "forms": './src/forms/index.jsx',
-    "idea-details": './src/idea-details/index.jsx',
-    "ideas-overview": './src/ideas-overview/index.jsx',
-    "idea-image": './src/image/index.jsx',
-    "ideas-on-map": './src/ideas-on-map/index.jsx',
-    "lightbox": './src/lightbox/index.jsx',
-    "modal-popup": './src/modal-popup/index.jsx',
-    "poll": './src/poll/index.jsx',
-    "previous-next-button-block": './src/previous-next-button-block/index.jsx',
-    "reactions": './src/reactions/index.jsx',
+    'all': './src/index.jsx',
+    'base-map': './src/base-map/index.jsx',
+    'button': './src/button/index.jsx',
+    'choices-guide': './src/choices-guide/index.jsx',
+    'forms': './src/forms/index.jsx',
+    'forms-image-upload': './src/forms/image-upload.jsx',
+    'idea-details': './src/idea-details/index.jsx',
+    'ideas-overview': './src/ideas-overview/index.jsx',
+    'idea-image': './src/image/index.jsx',
+    'ideas-on-map': './src/ideas-on-map/index.jsx',
+    'lightbox': './src/lightbox/index.jsx',
+    'modal-popup': './src/modal-popup/index.jsx',
+    'poll': './src/poll/index.jsx',
+    'previous-next-button-block': './src/previous-next-button-block/index.jsx',
+    'reactions': './src/reactions/index.jsx',
   },
 
 	output: {
@@ -38,7 +38,11 @@ module.exports = {
     'react-dom': 'ReactDOM',
 	},
 
+  devtool: false,
   plugins: [
+    new webpack.SourceMapDevToolPlugin({
+       filename: '[file].map',
+    }),
     new MiniCssExtractPlugin({
       filename: 'css/[name].css',
       ignoreOrder: false,
@@ -58,14 +62,14 @@ module.exports = {
 
 			{
 				test: /\.json$/,
-				loader: "json-loader"
+				loader: 'json-loader'
 			},
 
 			{
         test: /\.jsx?$/,
         exclude: /node_modules(?!\/openstad-component)/,
         use: {
-          loader: "babel-loader"
+          loader: 'babel-loader'
         }
 			},
 
@@ -92,19 +96,32 @@ module.exports = {
               publicPath: '../'
             }
           },
+          'css-loader',
           {
-            loader: "css-loader"
-          }
+            loader: 'postcss-loader',
+            options: {
+              postcssOptions: {
+                plugins: [
+                  [
+                    'postcss-preset-env',
+                    {
+                      // Options
+                    },
+                  ],
+                ],
+              },
+            },
+          },
         ]
       },
 
-      {
+      { // other images
         test: /\.(png|jpe?g|gif|svg)$/i,
         use: [
           {
-            loader: "file-loader",
+            loader: 'file-loader',
             options: {
-              name: "images/[name].[ext]",
+              name: 'images/[name].[ext]',
               publicPath: '../'
             },
           },